Fix incorrect rename in 0015b7281f9530088d604b6a0eefeae389c87aa9
